pyproject.toml
[tool.poetry]
name = "taikador-bot"
version = "0.1.0"
description = "A discord bot including lol features"
authors = ["Noah Parzentny <78926924+Taikador@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.2"
coloredlogs = "^15.0"
riotwatcher = "^3.1.1"
leaguenames = "^0.0.1"
tinydb = "^4.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
